+++ CHANGELOG.md
@@ -7,11 +7,14 @@ This is a patch / minor features release.
 ### New with 0.5.1
 
  * format.MetadataCopy has been updated to support free copying between metadata and payload
+ * producer.ElasticSearch alles setting the format of timeBasedIndex
 
 ### Fixed with 0.5.1
 
  * fixed inversion of -lc always
  * fixed a nil pointer panic with producer.elasticsearch when receiving messages with unassigned streams
+ * producer.ElasticSearch settings are now named according to config
+ * producer.ElasticSearch dayBasedIndex renamed to timeBasedIndex and it's now working as expected
 
 ## 0.5.0