+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This is a patch / minor features release.
  * format.MetadataCopy has been updated to support free copying between metadata and payload
  * producer.ElasticSearch alles setting the format of timeBasedIndex
  * format.GrokToJSON has new options: RemoveEmptyValues, NamedCapturesOnly and SkipDefaultPatterns
+ * using dep for dependencies instead of glide
 
 ### Fixed with 0.5.1
 
@@ -16,6 +17,7 @@ This is a patch / minor features release.
  * fixed a nil pointer panic with producer.elasticsearch when receiving messages with unassigned streams
  * producer.ElasticSearch settings are now named according to config
  * producer.ElasticSearch dayBasedIndex renamed to timeBasedIndex and it's now working as expected
+ * updated dependencies to latest version (brings support for kafka 1.0)
 
 ## 0.5.0